requirements.txt
pyttsx3==2.90
python-docx==0.8.10
docx2pdf==0.1.7